Modified lua cfunction sdl.text to return a surface without padding
The intent of this PR is to enable accurate measurements of text surfaces. With accurate width, it will be possible to chain several texts from different surfaces to form words and sentences that does not look disjointed.

It is inherently difficult to measure the actual length of a text when drawn with a font. The result was that the lua function `sdl.text` would return a surface with additional padding on both the left and the right ends.

This change removes that padding, so the returned surface is the exact width of the text (+1).

Two functions have been added to `Surface`, `padl`, and `padr`.
These can be used to retrieve the extra padding that would have been, if wanted.

The method used to trim the empty space on both sides of the string is as follows:
- Draw the string in red.
- Check the color bytes for red, and readjust the width based on the results.
- Redraw the string with accurate measurements.

This change will cause some texts in the ITB mod loader to be different than before, for left and right aligned text. It doesn't look bad; but a pass to update certain text positions might be in order.