Added support to draw surfaces with arbitrary scale and color
Exposed `screen:blitRect` to lua
This method can be used in place of `screen:blit` to draw surfaces with a specified destination rectangle; scaling the surface to fit inside it.

Added a parameter `color` to `blitRect` which is multiplied with the texture of the surface. The static color `sdl.color.white` can be used as argument if no color multiplication is needed.

Modified texture parameter for textures to scale with nearest neighbour interpolation to preserve hard edges of pixels for rescaled surfaces.